Fixed a lot of bugs + made multithread + reformatted code

+++ Interface/Server_Interface/Server_Interface/Protocol.cs
@@ -16,7 +16,7 @@ namespace Server_Interface
         {
             string newString;
             newString = startCharToServer.ToString();
-            newString += code;
+            newString += (int)code;
 
             foreach (string mess in value)
             {
